beter tests

+++ lib/gitdocs/migration/001_create_shares.rb
@@ -2,7 +2,7 @@ class CreateShares < ActiveRecord::Migration
   def self.up
     create_table :shares do |t|
       t.column :path, :string
-      t.column :polling_interval, :integer, :default => 15
+      t.column :polling_interval, :double, :default => 15
       t.column :notification, :boolean, :default => true
     end
   end
